+++ BallsVim/lua/plugins/old/core/ui.lua
@@ -1,228 +1,4 @@
--- from LazyVim
----@param component any
----@param text string
----@param hl_group? string
----@return string
-local function format(component, text, hl_group)
-  text = text:gsub("%%", "%%%%")
-  if not hl_group or hl_group == "" then
-    return text
-  end
-  ---@type table<string, string>
-  component.hl_cache = component.hl_cache or {}
-  local lualine_hl_group = component.hl_cache[hl_group]
-  if not lualine_hl_group then
-    local utils = require("lualine.utils.utils")
-    ---@type string[]
-    local gui = vim.tbl_filter(function(x)
-      return x
-    end, {
-      utils.extract_highlight_colors(hl_group, "bold") and "bold",
-      utils.extract_highlight_colors(hl_group, "italic") and "italic",
-    })
-
-    lualine_hl_group = component:create_hl({
-      fg = utils.extract_highlight_colors(hl_group, "fg"),
-      gui = #gui > 0 and table.concat(gui, ",") or nil,
-    }, "LV_" .. hl_group) --[[@as string]]
-    component.hl_cache[hl_group] = lualine_hl_group
-  end
-  return component:format_hl(lualine_hl_group) .. text .. component:get_default_hl()
-end
-
----@param opts? { modified_hl: string?, directory_hl: string?, filename_hl: string?, modified_sign: string?, readonly_icon: string?, length: number?}
-local function pretty_path(opts)
-  opts = vim.tbl_extend("force", {
-    modified_hl = "MatchParen",
-    directory_hl = "",
-    filename_hl = "Bold",
-    modified_sign = "",
-    readonly_icon = " 󰌾 ",
-    length = 3,
-  }, opts or {})
-
-  return function(self)
-    local path = vim.fn.expand("%:p") --[[@as string]]
-
-    if path == "" then
-      return ""
-    end
-
-    -- path = LazyVim.norm(path)
-    -- local root = LazyVim.root.get({ normalize = true })
-    -- local cwd = LazyVim.root.cwd()
-    local cwd = vim.fn.getcwd()
-
-    path = path:sub(#cwd + 2)
-
-    local sep = package.config:sub(1, 1)
-    local parts = vim.split(path, "[\\/]")
-
-    if opts.length == 0 then
-      parts = parts
-    elseif #parts > opts.length then
-      parts = { parts[1], "…", unpack(parts, #parts - opts.length + 2, #parts) }
-    end
-
-    if opts.modified_hl and vim.bo.modified then
-      parts[#parts] = parts[#parts] .. opts.modified_sign
-      parts[#parts] = format(self, parts[#parts], opts.modified_hl)
-    else
-      parts[#parts] = format(self, parts[#parts], opts.filename_hl)
-    end
-
-    local dir = ""
-    if #parts > 1 then
-      dir = table.concat({ unpack(parts, 1, #parts - 1) }, sep)
-      dir = format(self, dir .. sep, opts.directory_hl)
-    end
-
-    local readonly = ""
-    if vim.bo.readonly then
-      readonly = format(self, opts.readonly_icon, opts.modified_hl)
-    end
-    return dir .. parts[#parts] .. readonly
-  end
-end
-
 return {
-  {
-    "lualine.nvim",
-    for_cat = "core.ui",
-    event = "DeferredUIEnter",
-    after = function(_)
-      require('lualine').setup({
-        options = {
-          theme = "auto",
-          globalstatus = vim.o.laststatus == 3,
-          disabled_filetypes = { statusline = { "snacks_dashboard" } },
-        },
-        sections = {
-          lualine_a = { "mode" },
-          lualine_b = { "branch" },
-
-          lualine_c = {
-            vim.fn.getcwd(),
-            {
-              "diagnostics",
-              symbols = {
-                error = " ",
-                warn = " ",
-                info = " ",
-                hint = " ",
-              },
-            },
-            { "filetype", icon_only = true, separator = "", padding = { left = 1, right = 0 } },
-            { pretty_path() },
-          },
-          lualine_x = {
-            Snacks.profiler.status(),
-            -- stylua: ignore
-            {
-              function() return require("noice").api.status.command.get() end,
-              cond = function() return package.loaded["noice"] and require("noice").api.status.command.has() end,
-              color = function() return { fg = Snacks.util.color("Statement") } end,
-            },
-            -- stylua: ignore
-            {
-              function() return require("noice").api.status.mode.get() end,
-              cond = function() return package.loaded["noice"] and require("noice").api.status.mode.has() end,
-              color = function() return { fg = Snacks.util.color("Constant") } end,
-            },
-            -- stylua: ignore
-            {
-              function() return "  " .. require("dap").status() end,
-              cond = function() return package.loaded["dap"] and require("dap").status() ~= "" end,
-              color = function() return { fg = Snacks.util.color("Debug") } end,
-            },
-            -- stylua: ignore
-            {
-              "diff",
-              symbols = {
-                added = " ",
-                modified = " ",
-                removed = " ",
-              },
-              source = function()
-                local gitsigns = vim.b.gitsigns_status_dict
-                if gitsigns then
-                  return {
-                    added = gitsigns.added,
-                    modified = gitsigns.changed,
-                    removed = gitsigns.removed,
-                  }
-                end
-              end,
-            },
-          },
-          lualine_y = {
-            { "progress", separator = " ", padding = { left = 1, right = 0 } },
-            { "location", padding = { left = 0, right = 1 } },
-          },
-          lualine_z = {
-            function()
-              return " " .. os.date("%R")
-            end,
-          },
-        },
-      })
-    end,
-  },
-  {
-    "bufferline.nvim",
-    for_cat = "core.ui",
-    event = "DeferredUIEnter",
-    keys = {
-      { "<leader>bp", "<Cmd>BufferLineTogglePin<CR>", desc = "Toggle Pin" },
-      { "<leader>bP", "<Cmd>BufferLineGroupClose ungrouped<CR>", desc = "Delete Non-Pinned Buffers" },
-      { "<leader>br", "<Cmd>BufferLineCloseRight<CR>", desc = "Delete Buffers to the Right" },
-      { "<leader>bl", "<Cmd>BufferLineCloseLeft<CR>", desc = "Delete Buffers to the Left" },
-      { "<S-h>", "<cmd>BufferLineCyclePrev<cr>", desc = "Prev Buffer" },
-      { "<S-l>", "<cmd>BufferLineCycleNext<cr>", desc = "Next Buffer" },
-      { "[b", "<cmd>BufferLineCyclePrev<cr>", desc = "Prev Buffer" },
-      { "]b", "<cmd>BufferLineCycleNext<cr>", desc = "Next Buffer" },
-      { "[B", "<cmd>BufferLineMovePrev<cr>", desc = "Move buffer prev" },
-      { "]B", "<cmd>BufferLineMoveNext<cr>", desc = "Move buffer next" },
-    },
-    after = function(_)
-      require('bufferline').setup({
-        options = {
-          -- stylua: ignore
-          close_command = function(n) Snacks.bufdelete(n) end,
-          -- stylua: ignore
-          right_mouse_command = function(n) Snacks.bufdelete(n) end,
-          diagnostics = "nvim_lsp",
-          always_show_bufferline = false,
-          diagnostics_indicator = function(_, _, diag)
-            local icons = {
-              Error = " ",
-              Warn  = " ",
-              Hint  = " ",
-              Info  = " ",
-            }
-            local ret = (diag.error and icons.Error .. diag.error .. " " or "")
-              .. (diag.warning and icons.Warn .. diag.warning or "")
-            return vim.trim(ret)
-          end,
-          offsets = {
-            {
-              filetype = "neo-tree",
-              text = "Neo-tree",
-              highlight = "Directory",
-              text_align = "left",
-            },
-            {
-              filetype = "snacks_layout_box",
-            },
-          },
-          get_element_icon = function(element)
-            local icon, hl = require('nvim-web-devicons').get_icon_by_filetype(element.filetype, { default = false })
-            return icon, hl
-          end,
-        },
-      })
-    end,
-  },
   {
     "gitsigns.nvim",
     for_cat = "core.ui",
