+++ BallsVim/lua/specs/init.lua
@@ -4,4 +4,7 @@ return {
   { import = "specs.lsp" },
   { import = "specs.ui.which-key" },
   { import = "specs.ui.snacks" },
+  { import = "specs.ui.lualine" },
+  { import = "specs.ui.bufferline" },
+  { import = "specs.ui.web-devicons" },
 }
