+++ BallsVim/lua/specs/ui/web-devicons.lua
@@ -1,0 +1,6 @@
+return {
+  {
+    "nvim-web-devicons",
+    on_require = "nvim-web-devicons",
+  },
+}
